FIFO implementado
1) Uma classe apenas para entrada e saída (somente entrada por enquanto)
2) Adição de classes para os algoritmos (por enquanto, FIFO implementado)
3) Remoção da Main do código de leitura do arquivo, devido a 1).
4) Correção de algumas saídas "inesperadas"  em alguns métodos

+++ src/classes_basicas/ConfiguracoesMemoria.java
@@ -25,7 +25,7 @@ public class ConfiguracoesMemoria {
 	}
 	
 	private int calcularQtdPaginas() {
-		return Math.round(this.tamanhoMemoria/this.tamanhoPagina);
+		return (int) (Math.round((double)(this.tamanhoMemoria)/this.tamanhoPagina));
 	}
 	
 	public int getTamanhoMemoria() {
